working on compiling 1 qubit unitary

+++ utilities/database/database.py
@@ -39,7 +39,7 @@ def update_circuit_db_param_values(translator, circuit_db,symbol_to_value):
     trianables = circuit_db[circuit_db["trainable"] == True]
     trainable_symbols = trianables[~trianables["symbol"].isna()]
     if len(trainable_symbols) == len(symbol_to_value.values()):
-        circuit_db["param_value"].update({ind:val for ind, val in zip(trainable_symbols.index, symbol_to_value.values())})
+        circuit_db["param_value"].update({ind:val for ind, val in zip(trainable_symbols.index, symbol_to_value.values())}, dtype='float64')
     return circuit_db
 
 def give_resolver(translator, circuit_db):
